Extract event timeline source helpers
